change prevent_downgrade attribute to match resource
the attribute and the resource defaults disagree and that is confusing
and needs to be fixed.

this is the correct setting for configuration management.  to do config
management correctly you should be expressing your requirements as
equality pins.  if you have a later version installed you actually do
want to downgrade it.

otherwise you will not have consistent bugs across your production
environment, and your testing on any given version will be invalid.

to allow later versions on your systems means you are trusting that
i never write code which introduces bugs -- thank you for the vote of
confidence, but you're so very, very wrong.

this change will change behavior for some users.  perhaps surprsingly,
so we might already need to do a major bump to 3.0.0 here.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>