Make sure we actually see the upgrade converge by logging
If we kill/exec within the converge block we never get the message. Move that logic to its own block so we don't break why-run by still see the logging. I believe this 1/2 resolve some of the issues people have filed about logging and errors. They'll see the success now

Signed-off-by: Tim Smith <tsmith@chef.io>