Restart chef-client with cron under sysvinit
Normally we rely on a supervisor to restart chef-client after upgrade,
but in this case we can try to restart under sysvinit by using cron to
schedule the service to start again.

We need a reliable init script for the chef-client, so we check for
either the chef-client cookbook's presence and init_style or the
user explicitly configuring this pattern.

Signed-off-by: Bryan McLellan <btm@chef.io>