Replace hardcoded Chef strings with ChefUtils::Dist constants
This improves support for updating community distributions of Chef, such
as Cinc Client. Now when running under Cinc Client the proper paths and
product name will be be used.

Changes are backwards compatible with pre-ChefUtils clients and has no
net effect on Chef Infra Client as the hardcoded strings and constants
have the same values.

Signed-off-by: Joseph Larionov <jlarionov@webmd.net>