Improve Windows documentation
The documentation incorrectly recommended setting `post_install_action`
to `exec` on Windows when the code does not even allow that (it forces
the property to `kill` if given `exec` on Windows).

Update documentation for Windows to state that only
`kill` is supported for `post_install_action`. This has been the case
since 3.0.0.

Signed-off-by: Joseph Larionov <jlarionov@webmd.net>