Modifying mixlib-install gem 3.12.23
There is support available for the cinc in the chef_client_updater cookbook but for mixlib-install it was introduced with https://github.com/chef/mixlib-install/pull/340. 
The mixlib-install release is 3.12.23, which contains the changes to include cinc in the list of product matrix.

After updating found that the upgrade is still failing which is because chef_client_updater still contains the old release at https://github.com/chef-cookbooks/chef_client_updater/blob/main/providers/default.rb#L32-L37.